requirements.txt
trimesh==3.9.8
pyhocon==0.3.57
icecream==2.1.0
opencv_python==4.5.2.52
tqdm==4.50.2
scipy==1.7.0
numpy==1.19.2
pandas==1.4.0
PyMCubes==0.1.2
scikit-image
matplotlib
tensorboard
six
pillow
ninja
imageio
PyOpenGL
glfw
xatlas
gdown
Cython